Create improved profile follower command
Yes, this should be in ROOSTER, but it entails a level of encapsulation
that I don't want to deal with right now and bolting it on
to the existing robot code is a lot easier than writing a ton of test
boilerplate.

Basically, this follower is a complete rip-off of a method described by
254 on ChiefDelphi at https://www.chiefdelphi.com/forums/showthread.php?p=1702491#post1702491.
It lacks the bit where you add the velocity error to the commanded
velocity, because I'm fairly sure that it's only needed in a pure
pursuit controller setup. I don't want to write that because it involves
lots of math and I don't like math that much. So, we're using this
unless it's bad.

It does need a NavX to be bolted on at some point.