Set drivetrain velocity to 0 when profile ends
This fixes a very interesting bug where the profile would finish, but
due to the enclosing command group not ending, the PidDrive command
would not kick in. This caused the the drivetrain to continue to PID
(in position mode) to the last commanded position of the motion profile,
using coefficients for the velocity PID. Because of feed-forward, the
robot would then yeet itself across the room at max power.